we should rework the lemmings as faast as possible cause its the last thing we need to impliment and its the most important thing

+++ src/main/java/main/module/game/halloffame/HallOfFame.java
@@ -31,7 +31,7 @@ public class HallOfFame {
 		}
 	}
 
-	private boolean isHighScore (int score) {
+	private boolean isHighScore (long score) {
 		if (database.empty())
 			return true;
 		return database.peek().getScore() < score;
